+++ audionimbus/src/effect/direct.rs
@@ -201,6 +201,20 @@ impl Drop for DirectEffect {
 }
 
 unsafe impl Send for DirectEffect {}
+unsafe impl Sync for DirectEffect {}
+
+impl Clone for DirectEffect {
+    /// Retains an additional reference to the direct effect.
+    ///
+    /// The returned [`DirectEffect`] shares the same underlying Steam Audio object.
+    fn clone(&self) -> Self {
+        // SAFETY: The direct effect will not be destroyed until all references are released.
+        Self {
+            inner: unsafe { audionimbus_sys::iplDirectEffectRetain(self.inner) },
+            num_channels: self.num_channels,
+        }
+    }
+}
 
 /// Settings used to create a direct effect.
 #[derive(Debug)]
@@ -515,6 +529,27 @@ mod tests {
                     actual: 2,
                 })
             );
+        }
+    }
+
+    mod clone {
+        use super::*;
+
+        #[test]
+        fn test_clone() {
+            let context = Context::default();
+            let audio_settings = AudioSettings::default();
+
+            let effect = DirectEffect::try_new(
+                &context,
+                &audio_settings,
+                &DirectEffectSettings { num_channels: 1 },
+            )
+            .unwrap();
+            let clone = effect.clone();
+            assert_eq!(effect.raw_ptr(), clone.raw_ptr());
+            drop(effect);
+            assert!(!clone.raw_ptr().is_null());
         }
     }
 }